@@ -1,14 +1,26 @@
 {
   "name": "stellar-sdk",
-  "version": "0.12.0",
+  "version": "0.13.0",
   "description": "stellar-sdk is a library for working with the Stellar Horizon server.",
   "main": "lib/index.js",
   "scripts": {
     "test": "babel-node ./node_modules/.bin/gulp test",
-    "docs": "jsdoc ./src/ -d docs/ -r ./docs/readme.md",
+    "docs": "jsdoc -c .jsdoc.json --verbose",
     "preversion": "gulp test",
     "version": "gulp build",
-    "postversion": "git push && git push --tags"
+    "postversion": "git push && git push --tags",
+    "prettier-all": "prettier --write **/*.js"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "**/*.{js,json}": [
+      "prettier --write",
+      "git add"
+    ]
   },
   "keywords": [
     "stellar"
@@ -18,7 +30,7 @@
     "url": "https://github.com/stellar/js-stellar-sdk.git"
   },
   "engines": {
-    "node": ">=6"
+    "node": ">=6.14.0"
   },
   "author": "Stellar Development Foundation <hello@stellar.org>",
   "license": "Apache-2.0",
@@ -29,6 +41,7 @@
   "devDependencies": {
     "babel-cli": "^6.26.0",
     "babel-core": "~6.26.3",
+    "babel-eslint": "^10.0.1",
     "babel-istanbul": "^0.12.2",
     "babel-loader": "~6.3.2",
     "babel-plugin-transform-builtin-extend": "^1.1.2",
@@ -37,28 +50,36 @@
     "body-parser": "^1.12.2",
     "chai": "^2.2.0",
     "chai-as-promised": "^5.2.0",
+    "clear": "^0.1.0",
     "coveralls": "^2.11.4",
+    "eslint": "^5.12.1",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-config-prettier": "^3.6.0",
+    "eslint-plugin-import": "^2.15.0",
+    "eslint-plugin-node": "^8.0.1",
+    "eslint-plugin-prefer-import": "^0.0.1",
+    "eslint-plugin-prettier": "^3.0.1",
     "express": "^4.9.8",
     "ghooks": "^0.3.0",
     "gulp": "^3.8.11",
     "gulp-babel": "^6.1.3",
     "gulp-coveralls": "^0.1.3",
     "gulp-develop-server": "^0.4.2",
+    "gulp-eslint": "^5.0.0",
     "gulp-git": "~0.5.3",
     "gulp-insert": "^0.5.0",
     "gulp-istanbul": "^1.1.3",
-    "gulp-jshint": "^1.10.0",
     "gulp-load-plugins": "^0.9.0",
     "gulp-mocha": "^2.0.1",
     "gulp-plumber": "^1.0.0",
     "gulp-rename": "~1.2.0",
     "gulp-rimraf": "~0.1.0",
-    "gulp-uglify": "~0.3.1",
+    "gulp-uglify": "0.3.1",
     "gulp-webpack": "^1.3.1",
+    "husky": "^1.3.1",
     "isparta": "^4.0.0",
     "istanbul": "~0.3.2",
-    "jsdoc": "3.5.3",
-    "jshint-stylish": "^1.0.1",
+    "jsdoc": "3.5.5",
     "karma": "^0.13.2",
     "karma-chai": "^0.1.0",
     "karma-chai-as-promised": "^0.1.2",
@@ -70,9 +91,11 @@
     "karma-sinon": "^1.0.4",
     "karma-sinon-chai": "^0.3.0",
     "karma-webpack": "^1.7.0",
+    "lint-staged": "7.3.0",
     "lodash": "^4.17.10",
     "minami": "^1.1.1",
     "mocha": "~1.17.1",
+    "prettier": "^1.16.1",
     "run-sequence": "^1.0.2",
     "sinon": "^1.14.1",
     "sinon-chai": "^2.7.0",
@@ -84,7 +107,7 @@
     "event-source-polyfill": "0.0.12",
     "eventsource": "^1.0.5",
     "lodash": "^4.17.10",
-    "stellar-base": "^0.10.0",
+    "stellar-base": "^0.11.0",
     "toml": "^2.3.0",
     "urijs": "1.19.1"
   }
