@@ -1,17 +1,21 @@
-import {NotFoundError, NetworkError, BadRequestError} from "./errors";
+/*  */
+
 import forEach from 'lodash/forEach';
+import URI from 'urijs';
+import URITemplate from 'urijs/src/URITemplate';
+import axios from 'axios';
 
-let URI = require("urijs");
-let URITemplate = require("urijs/src/URITemplate");
+import { NotFoundError, NetworkError, BadRequestError } from './errors';
 
-let axios = require("axios");
-var EventSource = (typeof window === 'undefined') ? require('eventsource') : window.EventSource;
+const EventSource =
+  // eslint-disable-next-line no-undef, prefer-import/prefer-import-over-require
+  typeof window === 'undefined' ? require('eventsource') : window.EventSource;
 
 /**
  * Creates a new {@link CallBuilder} pointed to server defined by serverUrl.
  *
  * This is an **abstract** class. Do not create this object directly, use {@link Server} class.
- * @param {string} serverUrl
+ * @param {string} serverUrl URL of Horizon server
  * @class CallBuilder
  */
 export class CallBuilder {
@@ -23,28 +27,29 @@ export class CallBuilder {
 
   /**
    * @private
+   * @returns {void}
    */
   checkFilter() {
     if (this.filter.length >= 2) {
-      throw new BadRequestError("Too many filters specified", this.filter);
+      throw new BadRequestError('Too many filters specified', this.filter);
     }
 
     if (this.filter.length === 1) {
-      //append filters to original segments
-      let newSegment = this.originalSegments.concat(this.filter[0]);
+      // append filters to original segments
+      const newSegment = this.originalSegments.concat(this.filter[0]);
       this.url.segment(newSegment);
-    }        
+    }
   }
 
   /**
    * Triggers a HTTP request using this builder's current configuration.
-   * Returns a Promise that resolves to the server's response.
-   * @returns {Promise}
+   * @returns {Promise} a Promise that resolves to the server's response.
    */
   call() {
     this.checkFilter();
-    return this._sendNormalRequest(this.url)
-      .then(r => this._parseResponse(r));
+    return this._sendNormalRequest(this.url).then((r) =>
+      this._parseResponse(r)
+    );
   }
 
   /**
@@ -69,14 +74,14 @@ export class CallBuilder {
     // property is not reliable.
     let timeout;
 
-    var createTimeout = () => {
+    const createTimeout = () => {
       timeout = setTimeout(() => {
         es.close();
         es = createEventSource();
-      }, options.reconnectTimeout || 15*1000);
+      }, options.reconnectTimeout || 15 * 1000);
     };
 
-    var createEventSource = () => {
+    const createEventSource = () => {
       try {
         es = new EventSource(this.url.toString());
       } catch (err) {
@@ -89,17 +94,19 @@ export class CallBuilder {
 
       createTimeout();
 
-      es.onmessage = message => {
-        var result = message.data ? this._parseRecord(JSON.parse(message.data)) : message;
+      es.onmessage = (message) => {
+        const result = message.data
+          ? this._parseRecord(JSON.parse(message.data))
+          : message;
         if (result.paging_token) {
-          this.url.setQuery("cursor", result.paging_token);
+          this.url.setQuery('cursor', result.paging_token);
         }
         clearTimeout(timeout);
         createTimeout();
         options.onmessage(result);
       };
 
-      es.onerror = error => {
+      es.onerror = (error) => {
         if (options.onerror) {
           options.onerror(error);
         }
@@ -116,26 +123,34 @@ export class CallBuilder {
   }
 
   /**
+   * Convert a link object to a function that fetches that link.
    * @private
+   * @param {object} link A link object
+   * @param {bool} link.href the URI of the link
+   * @param {bool} [link.templated] Whether the link is templated
+   * @returns {function} A function that requests the link
    */
   _requestFnForLink(link) {
-    return opts => {
+    return (opts) => {
       let uri;
 
       if (link.templated) {
-        let template = URITemplate(link.href);
+        const template = URITemplate(link.href);
         uri = URI(template.expand(opts || {}));
       } else {
         uri = URI(link.href);
       }
 
-      return this._sendNormalRequest(uri).then(r => this._parseResponse(r));
+      return this._sendNormalRequest(uri).then((r) => this._parseResponse(r));
     };
-  } 
+  }
 
   /**
-   * Convert each link into a function on the response object.
+   * Given the json response, find and convert each link into a function that
+   * calls that link.
    * @private
+   * @param {object} json JSON response
+   * @returns {object} JSON response with string links replaced with functions
    */
   _parseRecord(json) {
     if (!json._links) {
@@ -143,15 +158,20 @@ export class CallBuilder {
     }
     forEach(json._links, (n, key) => {
       // If the key with the link name already exists, create a copy
-      if (typeof json[key] != 'undefined') {
+      if (typeof json[key] !== 'undefined') {
+        // eslint-disable-next-line no-param-reassign
         json[`${key}_attr`] = json[key];
       }
+
+      // eslint-disable-next-line no-param-reassign
       json[key] = this._requestFnForLink(n);
     });
     return json;
   }
-  
-  _sendNormalRequest(url) {
+
+  _sendNormalRequest(initialUrl) {
+    let url = initialUrl;
+
     if (url.authority() === '') {
       url = url.authority(this.url.authority());
     }
@@ -162,52 +182,63 @@ export class CallBuilder {
 
     // Temp fix for: https://github.com/stellar/js-stellar-sdk/issues/15
     url.setQuery('c', Math.random());
-    return axios.get(url.toString())
-      .then(response => response.data)
+    return axios
+      .get(url.toString())
+      .then((response) => response.data)
       .catch(this._handleNetworkError);
   }
 
   /**
    * @private
+   * @param {object} json Response object
+   * @returns {object} Extended response
    */
   _parseResponse(json) {
     if (json._embedded && json._embedded.records) {
       return this._toCollectionPage(json);
-    } else {
-      return this._parseRecord(json);
     }
+    return this._parseRecord(json);
   }
 
   /**
    * @private
+   * @param {object} json Response object
+   * @returns {object} Extended response object
    */
   _toCollectionPage(json) {
-    for (var i = 0; i < json._embedded.records.length; i++) {
+    for (let i = 0; i < json._embedded.records.length; i += 1) {
+      // eslint-disable-next-line no-param-reassign
       json._embedded.records[i] = this._parseRecord(json._embedded.records[i]);
     }
     return {
       records: json._embedded.records,
-      next: () => {
-        return this._sendNormalRequest(URI(json._links.next.href))
-          .then(r => this._toCollectionPage(r));
-      },
-      prev: () => {
-        return this._sendNormalRequest(URI(json._links.prev.href))
-          .then(r => this._toCollectionPage(r));
-      }
+      next: () =>
+        this._sendNormalRequest(URI(json._links.next.href)).then((r) =>
+          this._toCollectionPage(r)
+        ),
+      prev: () =>
+        this._sendNormalRequest(URI(json._links.prev.href)).then((r) =>
+          this._toCollectionPage(r)
+        )
     };
   }
 
   /**
    * @private
+   * @param {object} error Network error object
+   * @returns {Promise<Error>} Promise that rejects with a human-readable error
    */
   _handleNetworkError(error) {
     if (error.response && error.response.status) {
       switch (error.response.status) {
         case 404:
-          return Promise.reject(new NotFoundError(error.response.statusText, error.response.data));
+          return Promise.reject(
+            new NotFoundError(error.response.statusText, error.response.data)
+          );
         default:
-          return Promise.reject(new NetworkError(error.response.statusText, error.response.data));
+          return Promise.reject(
+            new NetworkError(error.response.statusText, error.response.data)
+          );
       }
     } else {
       return Promise.reject(new Error(error));
@@ -218,9 +249,10 @@ export class CallBuilder {
    * Sets `cursor` parameter for the current call. Returns the CallBuilder object on which this method has been called.
    * @see [Paging](https://www.stellar.org/developers/horizon/learn/paging.html)
    * @param {string} cursor A cursor is a value that points to a specific location in a collection of resources.
+   * @returns {object} current CallBuilder instance
    */
   cursor(cursor) {
-    this.url.setQuery("cursor", cursor);
+    this.url.setQuery('cursor', cursor);
     return this;
   }
 
@@ -228,18 +260,20 @@ export class CallBuilder {
    * Sets `limit` parameter for the current call. Returns the CallBuilder object on which this method has been called.
    * @see [Paging](https://www.stellar.org/developers/horizon/learn/paging.html)
    * @param {number} number Number of records the server should return.
+   * @returns {object} current CallBuilder instance
    */
   limit(number) {
-    this.url.setQuery("limit", number);
+    this.url.setQuery('limit', number);
     return this;
   }
 
   /**
    * Sets `order` parameter for the current call. Returns the CallBuilder object on which this method has been called.
-   * @param {"asc"|"desc"} direction
+   * @param {"asc"|"desc"} direction Sort direction
+   * @returns {object} current CallBuilder instance
    */
   order(direction) {
-    this.url.setQuery("order", direction);
+    this.url.setQuery('order', direction);
     return this;
   }
 }
