@@ -34,20 +34,16 @@ module.exports = function(config) {
       recordVideo: false
     },
 
-    browserDisconnectTimeout : 10000,
-    browserDisconnectTolerance : 1,
-    browserNoActivityTimeout : 4*60*1000,
-    captureTimeout : 4*60*1000,
+    browserDisconnectTimeout: 10000,
+    browserDisconnectTolerance: 1,
+    browserNoActivityTimeout: 4 * 60 * 1000,
+    captureTimeout: 4 * 60 * 1000,
 
     frameworks: ['mocha', 'chai-as-promised', 'chai', 'sinon'],
     customLaunchers: customLaunchers,
     browsers: Object.keys(customLaunchers),
 
-    files: [
-      'dist/stellar-sdk.js',
-      'test/test-helper.js',
-      'test/unit/**/*.js'
-    ],
+    files: ['dist/stellar-sdk.js', 'test/test-helper.js', 'test/unit/**/*.js'],
 
     preprocessors: {
       'test/**/*.js': ['webpack']
@@ -56,7 +52,7 @@ module.exports = function(config) {
     webpack: {
       module: {
         loaders: [
-          { test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader'}
+          { test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader' }
         ]
       }
     },
