@@ -1,13 +1,9 @@
 module.exports = function(config) {
   config.set({
     frameworks: ['mocha', 'chai-as-promised', 'chai', 'sinon'],
-    browsers : ["Firefox"],
+    browsers: ['Firefox'],
 
-    files: [
-      'dist/stellar-sdk.js',
-      'test/test-helper.js',
-      'test/unit/**/*.js'
-    ],
+    files: ['dist/stellar-sdk.js', 'test/test-helper.js', 'test/unit/**/*.js'],
 
     preprocessors: {
       'test/**/*.js': ['webpack']
@@ -16,7 +12,7 @@ module.exports = function(config) {
     webpack: {
       module: {
         loaders: [
-          { test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader'}
+          { test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader' }
         ]
       }
     },
