@@ -1,5 +1,5 @@
-import {Account as BaseAccount} from "stellar-base";
-import forIn from "lodash/forIn";
+import { Account as BaseAccount } from 'stellar-base';
+import forIn from 'lodash/forIn';
 
 /**
  * Do not create this object directly, use {@link Server#loadAccount}.
@@ -10,37 +10,39 @@ import forIn from "lodash/forIn";
  *
  * @see [Account Details](https://www.stellar.org/developers/horizon/reference/accounts-single.html)
  * @param {string} response Response from horizon account endpoint.
- * @returns {AccountResponse}
+ * @returns {AccountResponse} AccountResponse instance
  */
 export class AccountResponse {
-    constructor(response) {
-        this._baseAccount = new BaseAccount(response.account_id, response.sequence);
-        // Extract response fields
-        forIn(response, (value, key) => {
-            this[key] = value;
-        });
-    }
+  constructor(response) {
+    this._baseAccount = new BaseAccount(response.account_id, response.sequence);
+    // Extract response fields
+    forIn(response, (value, key) => {
+      this[key] = value;
+    });
+  }
 
-    /**
-     * Returns Stellar account ID, ex. `GB3KJPLFUYN5VL6R3GU3EGCGVCKFDSD7BEDX42HWG5BWFKB3KQGJJRMA`
-     * @returns {string}
-     */
-    accountId() {
-        return this._baseAccount.accountId();
-    }
+  /**
+   * Get Stellar account public key ex. `GB3KJPLFUYN5VL6R3GU3EGCGVCKFDSD7BEDX42HWG5BWFKB3KQGJJRMA`
+   * @returns {string} accountId
+   */
+  accountId() {
+    return this._baseAccount.accountId();
+  }
 
-    /**
-     * @returns {string}
-     */
-    sequenceNumber() {
-        return this._baseAccount.sequenceNumber();
-    }
+  /**
+   * Get the current sequence number
+   * @returns {string} sequenceNumber
+   */
+  sequenceNumber() {
+    return this._baseAccount.sequenceNumber();
+  }
 
-    /**
-     * Increments sequence number in this object by one.
-     */
-    incrementSequenceNumber() {
-        this._baseAccount.incrementSequenceNumber();
-        this.sequence = this._baseAccount.sequenceNumber();
-    }
+  /**
+   * Increments sequence number in this object by one.
+   * @returns {void}
+   */
+  incrementSequenceNumber() {
+    this._baseAccount.incrementSequenceNumber();
+    this.sequence = this._baseAccount.sequenceNumber();
+  }
 }
