@@ -1,5 +1,4 @@
-"use strict";
-
-module.exports = require("./index");
-module.exports.axios = require("axios");
-module.exports.StellarBase = require("stellar-base");
+/* eslint-disable prefer-import/prefer-import-over-require */
+module.exports = require('./index');
+module.exports.axios = require('axios');
+module.exports.StellarBase = require('stellar-base');
