@@ -1,4 +1,4 @@
-import {CallBuilder} from "./call_builder";
+import { CallBuilder } from './call_builder';
 
 /**
  * Creates a new {@link LedgerCallBuilder} pointed to server defined by serverUrl.
@@ -11,19 +11,18 @@ import {CallBuilder} from "./call_builder";
  * @param {string} serverUrl Horizon server URL.
  */
 export class LedgerCallBuilder extends CallBuilder {
-    constructor(serverUrl) {
-        super(serverUrl);
-        this.url.segment('ledgers');
-    }
+  constructor(serverUrl) {
+    super(serverUrl);
+    this.url.segment('ledgers');
+  }
 
-    /**
-     * Provides information on a single ledger.
-     * @param {number|string} sequence Ledger sequence
-     * @returns {LedgerCallBuilder}
-     */
-    ledger(sequence) {
-        this.filter.push(['ledgers', sequence.toString()]);
-        return this;
-    }
-
+  /**
+   * Provides information on a single ledger.
+   * @param {number|string} sequence Ledger sequence
+   * @returns {LedgerCallBuilder} current LedgerCallBuilder instance
+   */
+  ledger(sequence) {
+    this.filter.push(['ledgers', sequence.toString()]);
+    return this;
+  }
 }
