@@ -1,7 +1,14 @@
-import { CallBuilder } from "./call_builder";
-import { BadRequestError } from "./errors";
+import { CallBuilder } from './call_builder';
+import { BadRequestError } from './errors';
 
-const allowedResolutions = [60000, 300000, 900000, 3600000, 86400000, 604800000];
+const allowedResolutions = [
+  60000,
+  300000,
+  900000,
+  3600000,
+  86400000,
+  604800000
+];
 
 /**
  * Trade Aggregations facilitate efficient gathering of historical trade data.
@@ -19,67 +26,75 @@ const allowedResolutions = [60000, 300000, 900000, 3600000, 86400000, 604800000]
  * @param {long} offset segments can be offset using this parameter. Expressed in milliseconds. *Can only be used if the resolution is greater than 1 hour. Value must be in whole hours, less than the provided resolution, and less than 24 hours.
  */
 export class TradeAggregationCallBuilder extends CallBuilder {
-  constructor (serverUrl, base, counter, start_time, end_time, resolution, offset){
+  constructor(
+    serverUrl,
+    base,
+    counter,
+    start_time,
+    end_time,
+    resolution,
+    offset
+  ) {
     super(serverUrl);
-    
+
     this.url.segment('trade_aggregations');
     if (!base.isNative()) {
-        this.url.setQuery("base_asset_type", base.getAssetType());
-        this.url.setQuery("base_asset_code", base.getCode());
-        this.url.setQuery("base_asset_issuer", base.getIssuer());
+      this.url.setQuery('base_asset_type', base.getAssetType());
+      this.url.setQuery('base_asset_code', base.getCode());
+      this.url.setQuery('base_asset_issuer', base.getIssuer());
     } else {
-        this.url.setQuery("base_asset_type", 'native');
+      this.url.setQuery('base_asset_type', 'native');
     }
     if (!counter.isNative()) {
-        this.url.setQuery("counter_asset_type", counter.getAssetType());
-        this.url.setQuery("counter_asset_code", counter.getCode());
-        this.url.setQuery("counter_asset_issuer", counter.getIssuer());
+      this.url.setQuery('counter_asset_type', counter.getAssetType());
+      this.url.setQuery('counter_asset_code', counter.getCode());
+      this.url.setQuery('counter_asset_issuer', counter.getIssuer());
     } else {
-        this.url.setQuery("counter_asset_type", 'native');
+      this.url.setQuery('counter_asset_type', 'native');
     }
-    if ((typeof start_time === 'undefined') || (typeof end_time === 'undefined')) {
-        throw new BadRequestError("Invalid time bounds", [start_time, end_time]);
-    }else{
-        this.url.setQuery("start_time", start_time);
-        this.url.setQuery("end_time", end_time);
+    if (typeof start_time === 'undefined' || typeof end_time === 'undefined') {
+      throw new BadRequestError('Invalid time bounds', [start_time, end_time]);
+    } else {
+      this.url.setQuery('start_time', start_time);
+      this.url.setQuery('end_time', end_time);
     }
     if (!this.isValidResolution(resolution)) {
-        throw new BadRequestError("Invalid resolution", resolution);
-    }else{
-        this.url.setQuery("resolution", resolution);
+      throw new BadRequestError('Invalid resolution', resolution);
+    } else {
+      this.url.setQuery('resolution', resolution);
     }
     if (!this.isValidOffset(offset, resolution)) {
-        throw new BadRequestError("Invalid offset", offset);
+      throw new BadRequestError('Invalid offset', offset);
     } else {
-        this.url.setQuery("offset", offset);
+      this.url.setQuery('offset', offset);
     }
-
-
   }
 
   /**
    * @private
-   * @param {long} resolution 
+   * @param {long} resolution Trade data resolution in milliseconds
+   * @returns {boolean} true if the resolution is allowed
    */
-  isValidResolution(resolution){
+  isValidResolution(resolution) {
     let found = false;
 
-    for (let i = 0; i < allowedResolutions.length; i++) {
-        if (allowedResolutions[i] == resolution) {
-            found = true;
-            break;
-        }       
+    for (let i = 0; i < allowedResolutions.length; i += 1) {
+      if (allowedResolutions[i] === resolution) {
+        found = true;
+        break;
+      }
     }
     return found;
   }
 
-   /**
+  /**
    * @private
-   * @param {long} offset
-   * @param {long} resolution
+   * @param {long} offset Time offset in milliseconds
+   * @param {long} resolution Trade data resolution in milliseconds
+   * @returns {boolean} true if the offset is valid
    */
-  isValidOffset(offset, resolution){
+  isValidOffset(offset, resolution) {
     const hour = 3600000;
-    return !(offset > resolution || offset >= 24 * hour || (offset % hour !== 0));
+    return !(offset > resolution || offset >= 24 * hour || offset % hour !== 0);
   }
 }
