@@ -1,14 +1,14 @@
 if (typeof window === 'undefined') {
-  require("babel-register");
+  require('babel-register');
   global.StellarSdk = require('../src/index');
-  global.axios = require("axios");
-  var chaiAsPromised = require("chai-as-promised");
+  global.axios = require('axios');
+  var chaiAsPromised = require('chai-as-promised');
   global.chai = require('chai');
   global.chai.should();
   global.chai.use(chaiAsPromised);
   global.sinon = require('sinon');
-  global.Promise = require('bluebird');
   global.expect = global.chai.expect;
 } else {
+  // eslint-disable-next-line no-undef
   window.axios = StellarSdk.axios;
 }
